refactor: restructure project by organizing modules and updating routes; implement user, quiz, and result functionalities (#32)
