+++ src/app/http/router.js
@@ -1,7 +1,7 @@
-import authRoutes from "../../routes/authRoutes.js";
-import userRoutes from "../../routes/userRoutes.js";
-import quizRoutes from "../../routes/quizRoutes.js";
-import resultRoutes from "../../routes/resultRoutes.js";
+import authRoutes from "../../modules/auth/auth.routes.js";
+import userRoutes from "../../modules/users/user.routes.js";
+import quizRoutes from "../../modules/quizzes/quiz.routes.js";
+import resultRoutes from "../../modules/results/result.routes.js";
 
 export default async function router(app, option) {
     // Health check route
