+++ src/modules/results/result.routes.js
@@ -1,5 +1,5 @@
-import { getUserResults, saveResult, getResultById } from "../services/resultService.js";
-import { checkAuth } from "../shared/middleware/checkAuth.js";
+import { getUserResults, saveResult, getResultById } from "./result.service.js";
+import { checkAuth } from "../../shared/middleware/checkAuth.js";
 
 export default async function resultRoutes(fastify) {
 	fastify.addHook("preHandler", checkAuth);
